Expose commonly used Query methods in existing interface
There were few previous attempts to do so:
https://github.com/doctrine/orm/pull/1215
https://github.com/doctrine/orm/pull/6180
https://github.com/doctrine/orm/pull/6181

As well as complaints
https://github.com/doctrine/orm/discussions/10947

Removing final keyword has been rejected, as well as adding new interface.

This provides compromise: We don't remove final keyword nor add new interface. Instead, we add few methods to AbstractQuery. They are not required for classes extending AbstractQuery, hence no BC break there.
